requirements.txt
requests
matplotlib
schedule
pandas
ta
mplfinance
python-dotenv
